fix: data store error + id system

+++ src/main/java/boundary/observer/pembelian/PembelianEvent.java
@@ -2,14 +2,14 @@ package boundary.observer.pembelian;
 
 public class PembelianEvent {
   public String type, title, subtitle, imagePath;
-  public float price;
+  public double price;
   public int index, id;
 
   public PembelianEvent(String type) {
     this.type = type;
   }
 
-  public PembelianEvent(String type, int id, String title, String subtitle, float price, String imagePath) {
+  public PembelianEvent(String type, int id, String title, String subtitle, double price, String imagePath) {
     this.type = type;
     this.id = id;
     this.title = title;
