+++ src/main/java/util/RupiahConverter.java
@@ -4,7 +4,7 @@ import java.util.Locale;
 import java.text.NumberFormat;
 
 public class RupiahConverter {
-  public static String convert(float value) {
+  public static String convert(double value) {
     Locale locale = new Locale("in", "ID");
     NumberFormat formatter = NumberFormat.getCurrencyInstance(locale);
     return formatter.format(value);
